(wip) refactor selection routes
